+++ RELEASE_NOTES.md
@@ -16,6 +16,7 @@
 - Fix: User edit page does not save data when changing user role #60
 - Fix: Confusing UI response on instant new user account approval #64
 - Fix: Don't show alternative field in new user account approval mail, when user selected an institution
+- Fix: Make institution name available in new user approve mail
 
 
 ## Version 2023-02